Move healthcheck to routes directory
