+++ backend/app/src/routes/health.js
@@ -1,0 +1,9 @@
+'use strict'
+
+module.exports = async function (fastify, opts) {
+  fastify.route({
+    method: 'GET',
+    url: '/health',
+    handler: async (request, reply) => ({ status: 'pass' })
+  })
+}
