+++ backend/app/test/app.js
@@ -6,17 +6,3 @@ const fastify = buildFastify()
 test('app builds', async () => {
   await fastify.ready()
 })
-
-describe('GET /health', () => {
-  it('returns 200 with status in body', async () => {
-    const response = await fastify.inject({ // fastify.inject() will automatically wait until fastify is ready
-      method: 'GET',
-      path: '/health',
-      json: true
-    })
-    const body = response.json()
-
-    expect(response.statusCode).toBe(200)
-    expect(body.status).toBe('pass')
-  })
-})
